+++ crates/krb5-kdc/src/lib.rs
@@ -78,9 +78,10 @@ pub use store::{
     KDB_DISALLOW_PROXIABLE, KDB_DISALLOW_RENEWABLE, KDB_DISALLOW_SVR, KDB_DISALLOW_TGT_BASED,
     KDB_LOCKDOWN_KEYS, KDB_NO_AUTH_DATA_REQUIRED, KDB_OK_AS_DELEGATE, KDB_OK_TO_AUTH_AS_DELEGATE,
     KDB_PWCHANGE_SERVICE, KDB_REQUIRES_HW_AUTH, KDB_REQUIRES_PRE_AUTH, KDB_REQUIRES_PWCHANGE,
-    KDB_V1_BASE_LENGTH, KadmData, KeyEntry, MAX_ALIAS_DEPTH, NamedPolicy, Policy, Principal,
-    PrincipalStore, RID_ADMINISTRATOR, RID_FIRST_USER, RID_KRBTGT, S2K_ITERS, TlData, UlogEntry,
-    db_args_put_error, parse_spake_preauth_groups, random_key, s2k_params, strip_db_args,
+    KDB_V1_BASE_LENGTH, KadmData, KeyEntry, MAX_ALIAS_DEPTH, NamedPolicy, PWQUAL_DICT,
+    PWQUAL_EMPTY, PWQUAL_PRINC, Policy, Principal, PrincipalStore, RID_ADMINISTRATOR,
+    RID_FIRST_USER, RID_KRBTGT, S2K_ITERS, TlData, UlogEntry, db_args_put_error, parse_dict_words,
+    parse_spake_preauth_groups, random_key, s2k_params, strip_db_args,
 };
 
 use krb5_types::PrincipalName;